requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
amqp==5.0.9
    # via kombu
anyascii==0.2.0
    # via wagtail
asgiref==3.11.0
    # via
    #   django
    #   django-cors-headers
asttokens==2.2.1
    # via stack-data
backcall==0.1.0
    # via ipython
beautifulsoup4==4.8.2
    # via wagtail
billiard==3.6.4.0
    # via celery
boto3==1.16.63
    # via -r requirements.in
botocore==1.19.63
    # via
    #   boto3
    #   s3transfer
celery==5.2.2
    # via -r requirements.in
certifi==2026.1.4
    # via
    #   -r requirements.in
    #   opensearch-py
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==3.3.2
    # via
    #   pyopenssl
    #   social-auth-core
decorator==4.3.0
    # via ipython
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
    #   willow
dj-database-url==0.3.0
    # via -r requirements.in
django==5.2.9
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   django-hijack
    #   django-modelcluster
    #   django-permissionedforms
    #   django-redis
    #   django-role-permissions
    #   django-server-status
    #   django-storages
    #   django-stubs-ext
    #   django-taggit
    #   django-tasks
    #   django-treebeard
    #   djangorestframework
    #   jsonfield
    #   laces
    #   modelsearch
    #   social-auth-app-django
    #   wagtail
django-cors-headers==4.9.0
    # via -r requirements.in
django-discover-runner==1.0
    # via django-role-permissions
django-filter==25.2
    # via
    #   -r requirements.in
    #   wagtail
django-hijack==3.7.4
    # via -r requirements.in
django-ipware==7.0.1
    # via -r requirements.in
django-modelcluster==6.4.1
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-redis==5.2.0
    # via -r requirements.in
django-robots==6.1
    # via -r requirements.in
django-role-permissions==2.2.0
    # via -r requirements.in
django-server-status==0.7.3
    # via -r requirements.in
django-storages==1.14.6
    # via -r requirements.in
django-stubs-ext==5.2.8
    # via django-tasks
django-taggit==6.1.0
    # via
    #   -r requirements.in
    #   wagtail
django-tasks==0.9.0
    # via
    #   modelsearch
    #   wagtail
django-treebeard==4.8.0
    # via wagtail
django-webpack-loader==1.7.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   wagtail
draftjs-exporter==2.1.7
    # via wagtail
edx-api-client==1.2.0
    # via -r requirements.in
edx-opaque-keys==0.4
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
executing==1.2.0
    # via stack-data
factory-boy==2.8.1
    # via -r requirements.in
faker==0.9.1
    # via
    #   -r requirements.in
    #   factory-boy
filetype==1.2.0
    # via willow
flaky==3.7.0
    # via -r requirements.in
html5lib==1.1
    # via -r requirements.in
idna==3.7
    # via requests
ipython==8.10.0
    # via -r requirements.in
jedi==0.18.0
    # via ipython
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonfield==3.2.0
    # via -r requirements.in
jsonpatch==1.16
    # via -r requirements.in
jsonpointer==2.0
    # via jsonpatch
kombu==5.2.3
    # via celery
l18n==2021.3
    # via -r requirements.in
laces==0.1.2
    # via wagtail
matplotlib-inline==0.1.3
    # via ipython
modelsearch==1.1.1
    # via wagtail
natsort==8.4.0
    # via robohash
newrelic==6.0.1.155
    # via -r requirements.in
oauthlib==2.1.0
    # via
    #   requests-oauthlib
    #   social-auth-core
openpyxl==3.1.5
    # via wagtail
opensearch-dsl==1.0.0
    # via -r requirements.in
opensearch-py==1.0.0
    # via
    #   -r requirements.in
    #   opensearch-dsl
parso==0.8.1
    # via jedi
pbr==4.2.0
    # via stevedore
pexpect==4.6.0
    # via ipython
phonenumbers==9.0.21
    # via -r requirements.in
pickleshare==0.7.5
    # via ipython
pillow==12.1.0
    # via
    #   -r requirements.in
    #   pillow-heif
    #   robohash
    #   wagtail
pillow-heif==1.1.1
    # via willow
prompt-toolkit==3.0.36
    # via
    #   click-repl
    #   ipython
psycopg2==2.9.11
    # via -r requirements.in
ptyprocess==0.6.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycountry==16.11.27.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pygments==2.15.0
    # via ipython
pyjwt==2.8.0
    # via
    #   -r requirements.in
    #   open-discussions-client
    #   social-auth-core
pymongo==3.7.1
    # via edx-opaque-keys
pyopenssl==21.0.0
    # via -r requirements.in
python-dateutil==2.5.3
    # via
    #   -r requirements.in
    #   botocore
    #   edx-api-client
    #   faker
    #   opensearch-dsl
python-ipware==3.0.0
    # via django-ipware
python3-openid==3.1.0
    # via social-auth-core
pytz==2023.3
    # via
    #   celery
    #   l18n
redis==7.1.0
    # via
    #   -r requirements.in
    #   django-redis
requests==2.32.2
    # via
    #   -r requirements.in
    #   edx-api-client
    #   requests-oauthlib
    #   social-auth-core
    #   wagtail
requests-oauthlib==1.0.0
    # via social-auth-core
robohash==2.0
    # via -r requirements.in
s3transfer==0.3.7
    # via boto3
sentry-sdk==1.1.0
    # via -r requirements.in
six==1.14.0
    # via
    #   -r requirements.in
    #   asttokens
    #   click-repl
    #   cryptography
    #   django-role-permissions
    #   edx-opaque-keys
    #   faker
    #   html5lib
    #   l18n
    #   opensearch-dsl
    #   pyopenssl
    #   python-dateutil
    #   stevedore
social-auth-app-django==5.6.0
    # via -r requirements.in
social-auth-core==4.7.0
    # via
    #   -r requirements.in
    #   social-auth-app-django
soupsieve==2.0.1
    # via beautifulsoup4
sqlparse==0.5.5
    # via
    #   -r requirements.in
    #   django
stack-data==0.6.2
    # via ipython
stevedore==1.29.0
    # via edx-opaque-keys
telepath==0.3.1
    # via wagtail
text-unidecode==1.2
    # via faker
tornado==6.5.4
    # via -r requirements.in
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   django-stubs-ext
    #   django-tasks
urllib3==1.26.5
    # via
    #   -r requirements.in
    #   botocore
    #   opensearch-py
    #   requests
    #   sentry-sdk
uwsgi==2.0.22
    # via -r requirements.in
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==7.2.1
    # via -r requirements.in
wcwidth==0.1.7
    # via prompt-toolkit
webencodings==0.5.1
    # via html5lib
willow[heif]==1.12.0
    # via
    #   -r requirements.in
    #   wagtail

# The following packages are considered to be unsafe in a requirements file:
# setuptools
